--- src/main/java/baseball/utility/StringUtils.java
@@ -1,14 +1,0 @@
-package baseball.utility;
-
-import java.util.List;
-
-public class StringUtils {
-
-    public static List<Integer> convertStringToIntegerList(String string) {
-        return string.chars()
-                .map(character -> character - '0')
-                .boxed()
-                .toList();
-    }
-
-}
